Task 5 - Binance WebSocket Price Precision/binance-websocket-ticker/requirements.txt
websockets==12.0
python-dotenv==1.1.0
